* CRI-pips: (27 commits)
  use _int instead of intptr_t for local stacks intptr_t is a #define on bsd
  fix bugs in clean_unused_dynamic_variables formal parameters should not be cleaned
  some more PRIdPTR instead of %ld
  Remove all warnings from pips 64 bits main issues where related to usage of int where intptr_t should be used also note the existence of the c99 standard macro PRIdPTR that expands to ld or d for printf output of intptr_t
  Make svn ignore Makefile.in generated by autoconf
  undo pips_config.h.in removal this file is important for taking into account result of pips configuration
  Upgrade loop_expansion to use statement insertion Loop expansion can be used either as a stand alone (using guards) or as a three steps process: 1/ call loop_expansion_init to create the expanded statement 2/ call statement_insertion to verify the validity of the insertion 3/ call loop_expansion to that will perform the expansion, guarded or unguarded depending on the result of statement_insertion A step further for terapix code generation (and terapix paper)
  upgrade statement insertion it is now based on out regions for dead code detection, and uses pragma extension to output its result so that further processings are made easier
  Upgrade loop detection in controlizer, pattern i=i-k was not checked as a possible increment
  -Improving the treatment of arrays : p = tab[i] => (p, tab[i][0],Exact) -Implementing points_to_general_assignment() which treats every kind of assignment. When it's nt a basic case the result is an ANYWHERE points-to relation. -Factorizing tests on lhs and rhs at the level of one function: points_to_assignment. -Changing approximation of heap allocated data structure from MAY to EXACT. -More basic cases added as pointer to data structure's field. -Add comments on the function's organization and on old/new functions. -Start treating points-to relations at declarations, not yet fully implemented...
  Extension to expression instructions for some Feautrier's example; to be extended later by Molka Becher
  bug fix for Expressions/partial_eval02.c; to be checked better; comments added, reformatting,...
  In a previous commit, I fixed an error due to ECHO being defined twice this introduced a regression fixed here
  This is a big one: fix all -Wall warning on ia32 most warnings where caused by unused variables, no big bugs spotted there
  Finish the job BC began this morning for porting its change in autopips
  Add a new transformation to pips: statement insertion it checks if a pips inserted statement has little side effect and modify declaration accordingly. I am quite proud of it (although it a work in progress) for there are a lost of Psystemes inside :)
  Add a --enable-devel-mode in configure (default:no) it installs developer scripts (such as the one needed for validation) and sets CFLAGS to -g -O0 -Wall -Werror which is quite restrictive but should enforce good development practices :) as of now pips, linea and newgen compiles with thos flags on ia32
  add missing effects-util library to link, due to recent changes by BC.
  try to make also things ok for autotools. But it still fails.
  Libraries reorganization for effects
  ...

+++ packages/PIPS/pips/src/Libs/hpfc/defines-local.h
@@ -46,6 +46,7 @@
 #include "genC.h"
 #include "text.h"
 #include "ri.h"
+#include "effects.h"
 #include "database.h"
 #include "hpf.h"
 #include "hpf_private.h"
@@ -58,6 +59,7 @@
 #include "misc.h"
 #include "text-util.h"
 #include "ri-util.h"
+#include "effects-util.h"
 #include "hpfc.h"
 
 /* in paf-util.h:
@@ -145,8 +147,8 @@ void fprint_entity_list(FILE *fp, list l);
 
 #define what_stat_debug(level, stat)\
  ifdebug(level) \
- { int so_ = statement_ordering(stat);\
-   pips_debug(level, "statement %p (%d,%d:%d)\n",\
+ { intptr_t so_ = statement_ordering(stat);\
+   pips_debug(level, "statement %p (%"PRIdPTR",%"PRIdPTR":%"PRIdPTR")\n",\
    stat, ORDERING_NUMBER(so_), ORDERING_STATEMENT(so_), \
    statement_number(stat));}
